Fix temporary tarball being removed unnecesarilly

+++ beeflow/client/bee_client.py
@@ -173,6 +173,7 @@ def submit(wf_name: str = typer.Argument(..., help='The workflow name'),
            workdir: pathlib.Path = typer.Argument(...,
            help='working directory for workflow containing input + output files',)):
     """Submit a new workflow."""
+    tarball_path = ""
     if os.path.exists(wf_path):
         # Check to see if the wf_path is a tarball or a directory. Run package() if directory
         if os.path.isdir(wf_path):
@@ -224,7 +225,8 @@ def submit(wf_name: str = typer.Argument(..., help='The workflow name'),
     logging.info('Sumit workflow:  {resp.text}')
 
     # Cleanup code
-    os.remove(tarball_path)
+    if tarball_path:
+        os.remove(tarball_path)
     return wf_id
